Factor branded map into util
