+++ experimental/dds/tree2/src/core/index.ts
@@ -81,11 +81,7 @@ export {
 	PathRootPrefix,
 	isSkipMark,
 	emptyDelta,
-	AnchorKeyBrand,
 	AnchorSlot,
-	BrandedKey,
-	BrandedKeyContent,
-	BrandedMapSubset,
 	AnchorNode,
 	anchorSlot,
 	UpPathDefault,
