+++ experimental/dds/tree2/src/core/tree/index.ts
@@ -7,11 +7,7 @@ export {
 	Anchor,
 	AnchorLocator,
 	AnchorSet,
-	AnchorKeyBrand,
 	AnchorSlot,
-	BrandedKey,
-	BrandedKeyContent,
-	BrandedMapSubset,
 	AnchorNode,
 	anchorSlot,
 	AnchorEvents,
