+++ experimental/dds/tree2/src/internal.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-// Used by public types, but not part of the desired API surface
+// Used by public types, but not part of the desired API surface.
 export {
 	MakeNominal,
 	Invariant,
@@ -21,4 +21,5 @@ export {
 	AllowOptionalNotFlattened,
 	isAny,
 	RestrictiveReadonlyRecord,
+	BrandedKeyContent,
 } from "./util";
